- Update packages in embed and vue-app - Upgrade web3 - Upgrade inpage provider to the latest spec - Clean up code - bug fixes
