+++ examples/vue-app/package.json
@@ -9,20 +9,20 @@
   },
   "dependencies": {
     "@toruslabs/torus-embed": "^0.2.12",
-    "core-js": "^3.4.8",
+    "core-js": "^3.6.4",
     "eth-sig-util": "^2.5.2",
     "human-standard-token-abi": "^2.0.0",
-    "vue": "^2.6.10",
-    "web3": "^1.2.4"
+    "vue": "^2.6.11",
+    "web3": "^1.2.6"
   },
   "devDependencies": {
-    "@vue/cli-plugin-babel": "^4.1.1",
-    "@vue/cli-plugin-eslint": "^4.1.1",
-    "@vue/cli-service": "^4.1.1",
+    "@vue/cli-plugin-babel": "^4.1.2",
+    "@vue/cli-plugin-eslint": "^4.1.2",
+    "@vue/cli-service": "^4.1.2",
     "babel-eslint": "^10.0.3",
-    "eslint": "^6.7.2",
-    "eslint-plugin-vue": "^6.0.1",
-    "vue-template-compiler": "^2.6.10"
+    "eslint": "^6.8.0",
+    "eslint-plugin-vue": "^6.1.2",
+    "vue-template-compiler": "^2.6.11"
   },
   "eslintConfig": {
     "root": true,
