+++ src/inpage-provider.js
@@ -1,227 +1,272 @@
-import createErrorMiddleware from './createErrorMiddleware'
-import createTransformEthAddressMiddleware from './createTransformEthAddressMiddleware'
-import { setupMultiplex } from './stream-utils.js'
-import { Duplex as DuplexStream } from 'readable-stream'
-import { transformEthAddress } from './embedUtils'
-const pump = require('pump')
-const RpcEngine = require('json-rpc-engine')
-const createIdRemapMiddleware = require('json-rpc-engine/src/idRemapMiddleware')
-const createJsonRpcStream = require('json-rpc-middleware-stream')
-const SafeEventEmitter = require('safe-event-emitter')
-const log = require('loglevel')
-const ObservableStore = require('obs-store')
+import pump from 'pump'
+import RpcEngine from 'json-rpc-engine'
+import createIdRemapMiddleware from 'json-rpc-engine/src/idRemapMiddleware'
+import createJsonRpcStream from 'json-rpc-middleware-stream'
+import ObservableStore from 'obs-store'
+import asStream from 'obs-store/lib/asStream'
+import ObjectMultiplex from 'obj-multiplex'
+import SafeEventEmitter from 'safe-event-emitter'
+import { ethErrors } from 'eth-json-rpc-errors'
+import log from 'loglevel'
+import dequal from 'fast-deep-equal'
 
-class LocalStorageStream extends DuplexStream {
-  constructor(ethereum, web3) {
-    super({ objectMode: true })
-    this.ethereum = ethereum
-    this.web3 = web3
-  }
+import messages from './messages'
+// const { sendSiteMetadata } = require('./src/siteMetadata')
+import { createErrorMiddleware, logStreamDisconnectWarning } from './utils'
 
-  _read(chunk, enc, cb) {
-    log.info('reading from LocalStorageStore')
-  }
-
-  _onMessage(event) {
-    log.info('LocalStorageStore', event)
-  }
-
-  _write(chunk, enc, cb) {
-    const data = JSON.parse(chunk)
-    log.info('WRITING TO LOCALSTORAGESTREAM, CHUNK:', chunk)
-    for (const key in data) {
-      if (key === 'selectedAddress') {
-        if (data[key] !== '' && data[key] !== null && data[key] !== undefined) {
-          var prevSelectedAddress = window.sessionStorage.getItem('selectedAddress')
-          var newSelectedAddress = transformEthAddress(data[key])
-          this.web3.eth.defaultAccount = newSelectedAddress
-          this.ethereum.selectedAddress = newSelectedAddress
-          this.ethereum.publicConfigStore.updateState({ selectedAddress: newSelectedAddress })
-          window.sessionStorage.setItem('selectedAddress', newSelectedAddress)
-          if (prevSelectedAddress !== newSelectedAddress) {
-            this.emit('accountsChanged', [newSelectedAddress])
-          }
-        } else {
-          delete this.web3.eth.defaultAccount
-          delete this.ethereum.selectedAddress
-          window.sessionStorage.removeItem('selectedAddress')
-        }
-      } else if (key === 'networkVersion') {
-        window.sessionStorage.setItem(key, data[key])
-        if (this.ethereum.networkVersion !== data[key].toString()) {
-          this.ethereum.publicConfigStore.updateState({ networkVersion: data[key].toString() })
-          this.ethereum.networkVersion = data[key].toString()
-        }
-      } else {
-        window.sessionStorage.setItem(key, data[key])
-      }
-    }
-    cb()
-  }
+// resolve response.result, reject errors
+const getRpcPromiseCallback = (resolve, reject) => (error, response) => {
+  error || response.error ? reject(error || response.error) : Array.isArray(response) ? resolve(response) : resolve(response.result)
 }
 
 class MetamaskInpageProvider extends SafeEventEmitter {
   constructor(connectionStream) {
     super()
-    this.connectionStream = connectionStream
-    const self = this
-    self.isMetaMask = true
-    // super constructor
-    SafeEventEmitter.call(self)
-  }
+    this._state = {
+      sentWarnings: {
+        isConnected: false,
+        sendAsync: false,
+        sendSync: false
+      },
+      isConnected: undefined,
+      accounts: undefined,
+      isUnlocked: undefined
+    }
+    // public state
+    this.selectedAddress = null
+    this.networkVersion = undefined
+    this.chainId = undefined
 
-  init(opts = {}) {
-    const self = this
-    // subscribe to metamask public config (one-way)
     // setup connectionStream multiplexing
-    const mux = setupMultiplex(this.connectionStream)
-    const publicConfigStream = mux.createStream('publicConfig')
-    Object.defineProperty(self, 'mux', {
-      value: mux,
-      writable: true,
-      enumerable: false
-    })
+    const mux = (this.mux = new ObjectMultiplex())
+    pump(connectionStream, mux, connectionStream, this._handleDisconnect.bind(this, 'MetaMask'))
 
-    if (!opts.skipStatic) {
-      self.publicConfigStore = new ObservableStore({})
-      self.publicConfigStore.updateState = function(partialState) {
-        // if non-null object, merge
-        if (partialState && partialState instanceof Object) {
-          const state = this.getState()
-          const newState = Object.assign({}, state, partialState)
-          this.putState(newState)
-          // if not object, use new value
+    // subscribe to metamask public config (one-way)
+    this._publicConfigStore = new ObservableStore({ storageKey: 'MetaMask-Config' })
+
+    // handle isUnlocked changes, and chainChanged and networkChanged events
+    this._publicConfigStore.subscribe(stringifiedState => {
+      const state = JSON.parse(stringifiedState)
+      if ('isUnlocked' in state && state.isUnlocked !== this._state.isUnlocked) {
+        this._state.isUnlocked = state.isUnlocked
+        if (!this._state.isUnlocked) {
+          // accounts are never exposed when the extension is locked
+          this._handleAccountsChanged([])
         } else {
-          this.putState(partialState)
+          // this will get the exposed accounts, if any
+          try {
+            this._sendAsync(
+              { method: 'eth_accounts', params: [] },
+              () => {},
+              true // indicating that eth_accounts _should_ update accounts
+            )
+          } catch (_) {
+            // Swallow error
+          }
         }
       }
-      self.publicConfigStore.subscribe(function(state) {
-        opts.web3.eth.defaultAccount = state.selectedAddress
-      })
-    }
 
-    Object.defineProperty(self, 'lss', {
-      value: new LocalStorageStream(opts.ethereum, opts.web3),
-      writable: true,
-      enumerable: false
+      // Emit chainChanged event on chain change
+      if ('chainId' in state && state.chainId !== this.chainId) {
+        this.chainId = state.chainId
+        this.emit('chainChanged', this.chainId)
+      }
+
+      // Emit networkChanged event on network change
+      if ('networkVersion' in state && state.networkVersion !== this.networkVersion) {
+        this.networkVersion = state.networkVersion
+        this.emit('networkChanged', this.networkVersion)
+      }
     })
 
-    pump(publicConfigStream, self.lss)
+    pump(
+      mux.createStream('publicConfig'),
+      asStream(this._publicConfigStore),
+      // RPC requests should still work if only this stream fails
+      logStreamDisconnectWarning.bind(this, 'MetaMask PublicConfigStore')
+    )
+
     // ignore phishing warning message (handled elsewhere)
     mux.ignoreStream('phishing')
 
-    // connect to async provider
-    const jsonRpcConnection = createJsonRpcStream()
-    pump(jsonRpcConnection.stream, mux.getStream('provider'), jsonRpcConnection.stream, logStreamDisconnectWarning.bind(this, 'MetaMask RpcProvider'))
+    // setup own event listeners
 
-    // handle sendAsync requests via dapp-side rpc engine
-    const rpcEngine = new RpcEngine()
-    rpcEngine.push(createIdRemapMiddleware()) // TODO: fix metamask's janky way of keeping message ids unique
-    rpcEngine.push(createErrorMiddleware())
-    rpcEngine.push(createTransformEthAddressMiddleware())
-    rpcEngine.push(jsonRpcConnection.middleware)
-    self.rpcEngine = rpcEngine
-
-    // forward json rpc notifications
-    jsonRpcConnection.events.on('notification', payload => {
-      self.emit('data', null, payload)
+    // EIP-1193 connect
+    this.on('connect', () => {
+      this._state.isConnected = true
     })
 
-    // Work around for https://github.com/metamask/metamask-extension/issues/5459
-    // drizzle accidently breaking the `this` reference
-    self.send = self.send.bind(self)
-    self.sendAsync = self.sendAsync.bind(self)
+    // connect to async provider
+
+    const jsonRpcConnection = createJsonRpcStream()
+    pump(jsonRpcConnection.stream, mux.createStream('provider'), jsonRpcConnection.stream, this._handleDisconnect.bind(this, 'MetaMask RpcProvider'))
+
+    // handle RPC requests via dapp-side rpc engine
+    const rpcEngine = new RpcEngine()
+    rpcEngine.push(createIdRemapMiddleware())
+    rpcEngine.push(createErrorMiddleware())
+    rpcEngine.push(jsonRpcConnection.middleware)
+    this._rpcEngine = rpcEngine
+
+    // json rpc notification listener
+    // jsonRpcConnection.events.on('notification', payload => {
+    //   if (payload.method === 'wallet_accountsChanged') {
+    //     this._handleAccountsChanged(payload.result)
+    //   } else if (payload.method === 'eth_subscription') {
+    //     // EIP 1193 subscriptions, per eth-json-rpc-filters/subscriptionManager
+    //     this.emit('notification', payload.params.result)
+    //   }
+    // })
+
+    // send website metadata
+    // const domContentLoadedHandler = () => {
+    //   sendSiteMetadata(this._rpcEngine)
+    //   window.removeEventListener('DOMContentLoaded', domContentLoadedHandler)
+    // }
+    // window.addEventListener('DOMContentLoaded', domContentLoadedHandler)
+
+    // indicate that we've connected, for EIP-1193 compliance
+    setTimeout(() => this.emit('connect'))
   }
 
-  // Web3 1.0 provider uses `send` with a callback for async queries
-  send(payload, callback) {
-    const self = this
+  isMetaMask = true
 
-    if (callback) {
-      self.sendAsync(payload, callback)
-    } else {
-      return self._sendSync(payload)
+  /**
+   * Sends an RPC request to MetaMask. Resolves to the result of the method call.
+   * May reject with an error that must be caught by the caller.
+   *
+   * @param {(string|Object)} methodOrPayload - The method name, or the RPC request object.
+   * @param {Array<any>} [params] - If given a method name, the method's parameters.
+   * @returns {Promise<any>} - A promise resolving to the result of the method call.
+   */
+  send(methodOrPayload, params) {
+    // preserve original params for later error if necessary
+    const _params = params
+    let finalParams = params
+
+    // construct payload object
+    let payload
+    if (typeof methodOrPayload === 'object' && !Array.isArray(methodOrPayload)) {
+      // TODO:deprecate:2020-Q1
+      // handle send(object, callback), an alias for sendAsync(object, callback)
+      if (typeof finalParams === 'function') {
+        return this._sendAsync(methodOrPayload, finalParams)
+      }
+      payload = methodOrPayload
+    } else if (typeof methodOrPayload === 'string' && typeof finalParams !== 'function') {
+      // wrap params in array out of kindness
+      if (finalParams === undefined) {
+        finalParams = []
+      } else if (!Array.isArray(finalParams)) {
+        finalParams = [finalParams]
+      }
+
+      payload = {
+        method: methodOrPayload,
+        params: finalParams
+      }
     }
+
+    // typecheck payload and payload.method
+    if (Array.isArray(payload) || typeof finalParams === 'function' || typeof payload !== 'object' || typeof payload.method !== 'string') {
+      console.log(payload, 'error')
+      throw ethErrors.rpc.invalidRequest({
+        message: messages.errors.invalidParams(),
+        data: [methodOrPayload, _params]
+      })
+    }
+
+    return new Promise((resolve, reject) => {
+      try {
+        this._sendAsync(payload, getRpcPromiseCallback(resolve, reject))
+      } catch (error) {
+        reject(error)
+      }
+    })
   }
 
-  // handle sendAsync requests via asyncProvider
-  // also remap ids inbound and outbound
+  /**
+   * Deprecated.
+   * Backwards compatibility. ethereum.send() with callback.
+   *
+   * @param {Object} payload - The RPC request object.
+   * @param {Function} callback - The callback function.
+   */
   sendAsync(payload, cb) {
-    log.info('ASYNC REQUEST', payload)
-    const self = this
-    // fixes bug with web3 1.0 where send was being routed to sendAsync
-    // with an empty callback
-    if (cb === undefined) {
-      self.rpcEngine.handle(payload, noop)
-      return self._sendSync(payload)
-    } else {
-      self.rpcEngine.handle(payload, cb)
-    }
+    this._sendAsync(payload, cb)
   }
 
-  _sendSync(payload) {
-    const self = this
+  /**
+   * Internal RPC method. Forwards requests to background via the RPC engine.
+   * Also remap ids inbound and outbound.
+   *
+   * @param {Object} payload - The RPC request object.
+   * @param {Function} userCallback - The caller's callback.
+   * @param {boolean} isInternal - Whether the request is internal.
+   */
+  _sendAsync(payload, userCallback, isInternal = false) {
+    let cb = userCallback
 
-    let selectedAddress
-    let networkVersion
-    let result = null
-    switch (payload.method) {
-      case 'eth_accounts':
-        // read from localStorage
-        selectedAddress = window.sessionStorage.getItem('selectedAddress')
-        result = selectedAddress ? [selectedAddress] : []
-        break
+    if (!Array.isArray(payload)) {
+      if (!payload.jsonrpc) {
+        payload.jsonrpc = '2.0'
+      }
 
-      case 'eth_coinbase':
-        // read from localStorage
-        selectedAddress = window.sessionStorage.getItem('selectedAddress')
-        result = selectedAddress || null
-        break
-
-      case 'eth_uninstallFilter':
-        self.sendAsync(payload, noop)
-        result = true
-        break
-
-      case 'net_version':
-        log.info('NET VERSION REQUESTED')
-        networkVersion = window.sessionStorage.getItem('networkVersion')
-        result = networkVersion || null
-        break
-
-      // throw not-supported Error
-      default:
-        var link = 'https://github.com/MetaMask/faq/blob/master/DEVELOPERS.md#dizzy-all-async---think-of-metamask-as-a-light-client'
-        // eslint-disable-next-line max-len
-        var message = `The Torus Web3 object does not support synchronous methods like ${payload.method} without a callback parameter. See ${link} for details.`
-        throw new Error(message)
+      if (payload.method === 'eth_accounts' || payload.method === 'eth_requestAccounts') {
+        // handle accounts changing
+        cb = (err, res) => {
+          console.log(err, res, 'handling callback', payload)
+          this._handleAccountsChanged(res.result || [], payload.method === 'eth_accounts', isInternal)
+          userCallback(err, res)
+        }
+      }
     }
 
-    // return the result
-    return {
-      id: payload.id,
-      jsonrpc: payload.jsonrpc,
-      result: result
-    }
+    this._rpcEngine.handle(payload, cb)
   }
 
-  isConnected() {
-    return true
+  /**
+   * Called when connection is lost to critical streams.
+   */
+  _handleDisconnect(streamName, err) {
+    logStreamDisconnectWarning.bind(this)(streamName, err)
+    if (this._state.isConnected) {
+      this.emit('close', {
+        code: 1011,
+        reason: 'MetaMask background communication error.'
+      })
+    }
+    this._state.isConnected = false
+  }
+
+  /**
+   * Called when accounts may have changed.
+   */
+  _handleAccountsChanged(accounts, isEthAccounts = false, isInternal = false) {
+    // defensive programming
+    let finalAccounts = accounts
+    if (!Array.isArray(finalAccounts)) {
+      log.error('MetaMask: Received non-array accounts parameter. Please report this bug.', finalAccounts)
+      finalAccounts = []
+    }
+
+    // emit accountsChanged if anything about the accounts array has changed
+    if (!dequal(this._state.accounts, finalAccounts)) {
+      // we should always have the correct accounts even before eth_accounts
+      // returns, except in cases where isInternal is true
+      if (isEthAccounts && !isInternal) {
+        log.error('MetaMask: "eth_accounts" unexpectedly updated accounts. Please report this bug.', finalAccounts)
+      }
+
+      this.emit('accountsChanged', finalAccounts)
+      this._state.accounts = finalAccounts
+    }
+
+    // handle selectedAddress
+    if (this.selectedAddress !== finalAccounts[0]) {
+      this.selectedAddress = finalAccounts[0] || null
+    }
   }
 }
-
-// util
-
-function logStreamDisconnectWarning(remoteLabel, err) {
-  let warningMsg = `MetamaskInpageProvider - lost connection to ${remoteLabel}`
-  if (err) warningMsg += '\n' + err.stack
-  log.warn(warningMsg)
-  const listeners = this.listenerCount('error')
-  if (listeners > 0) {
-    this.emit('error', warningMsg)
-  }
-}
-
-function noop() {}
 
 export default MetamaskInpageProvider
