+++ src/stream-utils.js
@@ -1,36 +1,12 @@
-import Through from 'through2'
 import ObjectMultiplex from 'obj-multiplex'
 import pump from 'pump'
-
-/**
- * Returns a stream transform that parses JSON strings passing through
- * @return {stream.Transform}
- */
-const jsonParseStream = () => {
-  return Through.obj(function(serialized, _, cb) {
-    this.push(JSON.parse(serialized))
-    cb()
-  })
-}
-
-/**
- * Returns a stream transform that calls {@code JSON.stringify}
- * on objects passing through
- * @return {stream.Transform} the stream transform
- */
-const jsonStringifyStream = () => {
-  return Through.obj(function(obj, _, cb) {
-    this.push(JSON.stringify(obj))
-    cb()
-  })
-}
 
 /**
  * Sets up stream multiplexing for the given stream
  * @param {any} connectionStream - the stream to mux
  * @return {stream.Stream} the multiplexed stream
  */
-const setupMultiplex = connectionStream => {
+export const setupMultiplex = connectionStream => {
   const mux = new ObjectMultiplex()
   // bind helper method to get previously created streams
   mux.getStream = function(name) {
@@ -46,5 +22,3 @@ const setupMultiplex = connectionStream => {
   })
   return mux
 }
-
-export { jsonParseStream, jsonStringifyStream, setupMultiplex }
