+++ src/utils.js
@@ -1,3 +1,8 @@
+import log from 'loglevel'
+import { serializeError } from 'eth-json-rpc-errors'
+import EventEmitter from 'events'
+import SafeEventEmitter from 'safe-event-emitter'
+
 import config from './config'
 
 const paymentProviders = config.paymentProviders
@@ -27,4 +32,48 @@ export const validatePaymentProvider = (provider, params) => {
     errors.selectedCryptoCurrency = 'Unsupported cryptoCurrency'
 
   return { errors, isValid: Object.keys(errors).length === 0 }
+}
+
+/**
+ * Middleware configuration object
+ *
+ * @typedef {Object} MiddlewareConfig
+ */
+
+/**
+ * json-rpc-engine middleware that both logs standard and non-standard error
+ * messages and ends middleware stack traversal if an error is encountered
+ *
+ * @returns {Function} json-rpc-engine middleware function
+ */
+export const createErrorMiddleware = () => {
+  return (_req, res, next) => {
+    next(done => {
+      const { error } = res
+      if (!error) {
+        return done()
+      }
+      serializeError(error)
+      log.error(`MetaMask - RPC Error: ${error.message}`, error)
+      done()
+    })
+  }
+}
+
+/**
+ * Logs a stream disconnection error. Emits an 'error' if bound to an
+ * EventEmitter that has listeners for the 'error' event.
+ *
+ * @param {string} remoteLabel - The label of the disconnected stream.
+ * @param {Error} err - The associated error to log.
+ */
+export const logStreamDisconnectWarning = (remoteLabel, err) => {
+  let warningMsg = `MetamaskInpageProvider - lost connection to ${remoteLabel}`
+  if (err) warningMsg += '\n' + err.stack
+  log.warn(warningMsg)
+  if (this instanceof EventEmitter || this instanceof SafeEventEmitter) {
+    if (this.listenerCount('error') > 0) {
+      this.emit('error', warningMsg)
+    }
+  }
 }
